Added more tools for Twitter matching on candidate_edit and candidate_list. Added the ability to assign voter guide possibility entries to specific Political Data Managers.
