+++ import_export_twitter/urls.py
@@ -37,4 +37,10 @@ urlpatterns = [
     url(r'^transfer_candidate_twitter_handles_from_google_civic/$',
         views_admin.transfer_candidate_twitter_handles_from_google_civic_view,
         name='transfer_candidate_twitter_handles_from_google_civic'),
+    url(r'^delete_possible_twitter_handles/(?P<candidate_campaign_we_vote_id>wv[\w]{2}cand[\w]+)/$',
+        views_admin.delete_possible_twitter_handles_view, name='delete_possible_twitter_handles', ),
+    url(r'^retrieve_possible_twitter_handles/(?P<candidate_campaign_we_vote_id>wv[\w]{2}cand[\w]+)/$',
+        views_admin.retrieve_possible_twitter_handles_view, name='retrieve_possible_twitter_handles', ),
+    url(r'^bulk_retrieve_possible_twitter_handles/$',
+        views_admin.bulk_retrieve_possible_twitter_handles_view, name='bulk_retrieve_possible_twitter_handles', ),
 ]
