+++ templates/voter_guide/voter_guide_possibility_table.html
@@ -4,6 +4,7 @@
         <thead>
         <tr>
             <th></th>
+            <th>Assigned to</th>
             <th>Organization/Candidate</th>
             <th>Source</th>
             <th>Ballot Items</th>
@@ -16,6 +17,11 @@
     {% for voter_guide_possibility in voter_guide_possibility_list %}
         <tr>
             <td>{{ forloop.counter|add:starting_counter_number }}</td>
+            <td>
+              {% if voter_guide_possibility.assigned_to_name %}
+                {{ voter_guide_possibility.assigned_to_name }}
+              {% endif %}
+            </td>
             <td>
             {% if voter_guide_possibility.voter_guide_possibility_type == ORGANIZATION_ENDORSING_CANDIDATES or voter_guide_possibility.voter_guide_possibility_type == UNKNOWN_TYPE %}
                 {% if voter_guide_possibility.organization %}
