+++ twitter/models.py
@@ -220,13 +220,48 @@ class TwitterUserManager(models.Model):
         }
         return results
 
-    def update_or_create_twitter_link_possibility(self, candidate_campaign_we_vote_id, twitter_json, search_term,
-                                                  likelihood_score):
+    def update_or_create_twitter_link_possibility(
+            self,
+            twitter_link_possibility_id=0,
+            candidate_campaign_we_vote_id='',
+            twitter_handle='',
+            defaults={}):
+        status = ""
+        try:
+            if positive_value_exists(twitter_link_possibility_id):
+                twitter_link_possibility = TwitterLinkPossibility.objects.get(id=twitter_link_possibility_id)
+                change_to_save = False
+                if 'not_a_match' in defaults:
+                    change_to_save = True
+                    twitter_link_possibility.not_a_match = defaults['not_a_match']
+                if positive_value_exists(change_to_save):
+                    twitter_link_possibility.save()
+            else:
+                TwitterLinkPossibility.objects.update_or_create(
+                    candidate_campaign_we_vote_id=candidate_campaign_we_vote_id,
+                    twitter_handle__iexact=twitter_handle,
+                    defaults=defaults,
+                    )
+            status += "TWITTER_LINK_TO_POSSIBILITY_UPDATE_OR_CREATED "
+            success = True
+
+        except Exception as e:
+            status += "TWITTER_LINK_TO_POSSIBILITY_NOT_UPDATE_OR_CREATED " + str(e) + ' '
+            success = False
+
+        results = {
+            'success': success,
+            'status': status,
+        }
+        return results
+
+    def update_or_create_twitter_link_possibility_from_twitter_json(
+            self, candidate_campaign_we_vote_id, twitter_json, search_term, likelihood_score):
         try:
             TwitterLinkPossibility.objects.update_or_create(
                 candidate_campaign_we_vote_id=candidate_campaign_we_vote_id,
                 twitter_id=twitter_json['id'],
-                twitter_handle=twitter_json['screen_name'],
+                twitter_handle__iexact=twitter_json['screen_name'],
                 defaults={
                     'likelihood_score': likelihood_score,
                     'search_term_used': search_term,
